Add a daily barchart
THis adds a bar chart showing the number of appication creations over the past 30 days.

There is also an endpoint to return the data as Json.

Both are behind key based authentication.

The key based auth has been extended to allow the key to be specified as a query param.

Closes #13